Splitting up query config files and tests
